+++ src/bancoimobiliario/GameConst.java
@@ -1,0 +1,8 @@
+package bancoimobiliario;
+
+public class GameConst {
+    public int saldoInicial;
+    public int salarioPorVolta;
+    public int maxRodadas;
+    public int rodadaAtual;
+}
